Http client

+++ internal/model/payment.go
@@ -5,7 +5,7 @@ import "time"
 type Payment struct {
 	Id       int       `json:"id" db:"id"`
 	Person   string    `json:"person" db:"person"`
-	Amount   int       `json:"amount" db:"amount"`
+	Amount   float64   `json:"amount" db:"amount"`
 	Currency string    `json:"currency" db:"currency"`
 	Time     time.Time `json:"time" db:"time"`
 }
